Retyped price to double

+++ src/main/java/org/example/client/models/Product.java
@@ -8,7 +8,7 @@ public class Product {
     private String name;
     private String brand;
     private Integer packageSize;
-    private Double price;
+    private Float price;
     private String image;
     private Integer availability;
     private String characteristics;
@@ -16,7 +16,7 @@ public class Product {
     // For shopping cart
     private Integer quantity = 0;
 
-    public Product(Integer id, String name, String brand, Integer packageSize, Double price, String image,
+    public Product(Integer id, String name, String brand, Integer packageSize, Float price, String image,
                    Integer availability, String characteristics, String section) {
         this.id = id;
         this.name = name;
@@ -34,7 +34,7 @@ public class Product {
         this.name = json.getString("name");
         this.brand = json.getString("brand");
         this.packageSize = json.getInt("package_size");
-        this.price = json.getDouble("price");
+        this.price = json.getFloat("price");
         if (json.has("image"))
             this.image = json.getString("image");
         else
@@ -72,11 +72,11 @@ public class Product {
         this.packageSize = packageSize;
     }
 
-    public Double getPrice() {
+    public Float getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Float price) {
         this.price = price;
     }
